Upgrade to JUnit 4.13 and clean up the tests a bit.

+++ artio-system-tests/src/test/java/uk/co/real_logic/artio/system_tests/GatewayToGatewaySystemTest.java
@@ -35,6 +35,7 @@ import uk.co.real_logic.artio.session.Session;
 import java.util.List;
 import java.util.function.IntSupplier;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.*;
 import static uk.co.real_logic.artio.Constants.*;
@@ -738,7 +739,7 @@ public class GatewayToGatewaySystemTest extends AbstractGatewayToGatewaySystemTe
 
         assertInitSeqNum(2, 2, 0);
 
-        final Reply<?> resetSequenceNumber = resetSequenceNumber((long)400);
+        final Reply<?> resetSequenceNumber = resetSequenceNumber(400);
         assertTrue("Should have errored: " + resetSequenceNumber, resetSequenceNumber.hasErrored());
         final String message = resetSequenceNumber.error().getMessage();
         assertTrue(message, message.contains("Unknown sessionId: 400"));
